pyproject.toml
[project]
name = "glm-bigmeg"
version = "1.0.0"
description = "Analysis code for Robust and replicable effects of ageing on resting state brain electrophysiology measured with MEG"
readme = "README.md"
requires-python = "==3.11.3"
license = {text = "MIT"}
authors = [
    {name = "Andrew Quinn", email = "a.quinn@bham.ac.uk"},
]
dependencies = [
    "glmtools>=0.2.1",
    "ipython>=8.12.3",
    "isort>=6.0.1",
    "matplotlib==3.9.3",
    "matplotlib-label-lines>=0.8.0",
    "meegkit>=0.1.7",
    "numpy==2.0.2",
    "openpyxl>=3.1.5",
    "osl>=1.1.0",
    "pandas==2.2.2",
    "pymc-bart>=0.9.2",
    "pyside2>=5.15.2.1",
    "sails>=1.2.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "setuptools>=75.3.2",
]

[project.urls]
Homepage = ""
Preprint = ""

[tool.uv]
# https://docs.astral.sh/uv/reference/settings/#package
package = false
# https://docs.astral.sh/uv/reference/settings/#override-dependencies
override-dependencies = [
    "mne==1.8.0"
]
